Sort siblings by age Include refactor of `get_age` and `valid_month`
